% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Data.R
\docType{data}
\name{AA_Data}
\alias{AA_Data}
\title{Double-tag (AA) data in wide format}
\format{An object of class \code{data.frame} with 21 rows and 3 columns.}
\usage{
AA_Data
}
\description{
Double-tag (AA) data in wide format
}
\details{
As part of a Fisheries Research and Developement Corporation (FRDC) research project conducted by the Department of Primary Industries and Regional Developement (DPIRD) 600 Western Rock Lobsters were double-tagged with two identical T-bar anchor tags (see de Lestang et al., 2016 for details).  Time and location of all recaptures and the number of tags retained have been reported to DPIRD by commercial fishermen.  This data contains only the liberty time in months (column 1), and for each liberty time the number of lobsters recaptured with two or only one tags still present.
}
\examples{
AA_Data

}
\references{
de Lestang, S., Caputi, N. and How, J. 2016. Western Australian Marine Stewardship Council Report Series No. 9: Resource Assessment Report: Western Rock Lobster Resource of Western Australia. Department of Fisheries, Western Australia. http://www.fish.wa.gov.au/Documents/wamsc_reports/wamsc_report_no_9.pdf
}
\keyword{datasets}
