% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_Data.R
\docType{data}
\name{ab_Data}
\alias{ab_Data}
\title{Double-mark (ab) data in wide format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.}
\usage{
ab_Data
}
\description{
Double-mark (ab) data in wide format
}
\details{
As part of a Department of Primary Industries and Regional Developement (DPIRD) research project over 2000 Western Rock Lobsters were double marked: One T-bar tag was inserted ventrally (see de Lestang et al., 2016 for details) and a month-specific pleopod was snipped from the animal (the snipped pleopod provides a semi-permanent mark which is re-snipped upon subsequent recaptures therefore extending its mark life indefinately).  Time of all recaptures and the presence or absense of tags when snipped pleopods were identified were  reported by DPIRD staff.  This data contains only the liberty time in months (column 1), and for each liberty time the number of lobsters recaptured with both marks (tag and snipped pleopod), only a snipped ploepod (column 2) and only a tag (column 3).
}
\examples{
ab_Data

}
\references{
de Lestang, S., Caputi, N. and How, J. 2016. Western Australian Marine Stewardship Council Report Series No. 9: Resource Assessment Report: Western Rock Lobster Resource of Western Australia. Department of Fisheries, Western Australia. http://www.fish.wa.gov.au/Documents/wamsc_reports/wamsc_report_no_9.pdf
}
\keyword{datasets}
