% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAA.R
\name{PAA}
\alias{PAA}
\title{Function to return the probability of recapturing an animal with two of the same tags.}
\usage{
PAA(pin, t)
}
\arguments{
\item{pin}{A vector of input parameters (length two):
Instant tag retension (proportion)
Chronic tag loss (proportion)}

\item{t}{A vector of recapture liberties.}
}
\value{
The probability of two tags remaining after time t.
}
\description{
Used by the TagLoss model when animals have been tagged with two of the same tags which are expected to have the same probability of retension.
}
\examples{
PAA(c(0.5,0.1),1:10)
}
