% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pb1.R
\name{Pb}
\alias{Pb}
\title{Function used by the TagLoss model.}
\usage{
Pb(pin, t)
}
\arguments{
\item{pin}{A vector of input parameters (length four):
Instant tag retension of 'a' tag (proportion)
Chronic tag loss of 'a' tag (proportion)
Instant tag retension of 'b' tag (proportion)
Chronic tag loss of 'b' tag (proportion)}

\item{t}{A vector of recapture liberties.}
}
\value{
The probability of tag 'b' remaining after time t.
}
\description{
Used by the TagLoss model when animals have been tagged with two tags which are expected to have different probabilities of retension.
}
\examples{
Pb(c(0.5,0.1,1.0,0.2),1:10)
}
