% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMod.R
\name{PlotMod}
\alias{PlotMod}
\title{Plots the output from the TagLoss model}
\usage{
PlotMod(Data, col = c(2:4), boot = 0, resid = FALSE)
}
\arguments{
\item{Data}{The original input data used for the model.}

\item{col}{The colours to use for each tag type}

\item{boot}{The number of bootstraps used (if any).  This is used to detemine whether bootstrap results should be used.}

\item{resid}{A residual plot (observed - estimated  proportions).  Only used when boot is set to 0.}
}
\value{
Bar and residual plots.
}
\description{
This function plots the observed and estimated proportions of tags recaptured (and the residuals between them is requested).
}
\examples{
  Solve_TLmodel(c(0.5,0.1), AA_Data)
  PlotMod(AA_Data, col=c('green','purple', 'red'))
}
