% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solve_TLmodel.R
\name{Solve_TLmodel}
\alias{Solve_TLmodel}
\title{Uses nlminb optimisation to solve the TagLoss model}
\usage{
Solve_TLmodel(pin, Data, trace = FALSE, bound = TRUE, boot = 0,
  probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{pin}{The input parameters.  Takes a vector of two or four numbers.}

\item{Data}{Observed data as a data.frame in wide format.  rows should represent unique periods at liberty identified by the first column (liberty).  Subsequent column record the numbers of observation of each tag recapture (e.g. column two can be those recaptured with two tags and column three those with only one tag).}

\item{trace}{The value of the objective function and the parameters is printed every iteration. Defaults to 0 which indicates no trace information is to be printed.}

\item{bound}{Whether the parameters should be bounded betweeen 0 and 1.  Logical.}

\item{boot}{Whether bootstrapping should occur (occurs if > 1) and how many iterations should be used in the bootstrapping proceedure.}

\item{probs}{The percentiles used in summarising the bootstrapping proceedure (defaluts to 2.5, 50 and 95th percentiles)}
}
\value{
A list "Modelout" containing parameter estimates (back-transformed if bound=TRUE) and a summary of the optimisation process.  If boot >1 bootstrapping occurs and percentile estimates are generated for all parameters and the estimated numbers of tag recaptures in each group.
}
\description{
This function minimises the optimisation process prodced by the TagLoss model.
}
\examples{
Solve_TLmodel(c(0.5,0.1), AA_Data)
}
