% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagLoss.R
\name{TagLoss}
\alias{TagLoss}
\title{Models tag-loss rate}
\usage{
TagLoss(pin, bound = TRUE, obs = obs, t = t, flag = "solve")
}
\arguments{
\item{pin}{The input parameters.  Take a vector of two numbers.}

\item{bound}{Whether the parameters should be bounded betweeen 0 and 1.  A logical.}

\item{obs}{The observed recaptures of tagged animals.  A matrix with recapture state as rows and observations from different time periods as columns.}

\item{t}{Time periods. A numeric vector.}

\item{flag}{Switches the model to return the Log-likelihood or a matrix of estimated probabilities}
}
\value{
A list containing parameter estimates (back-transformed if bound=TRUE) and a summary of the optimisation process.
}
\description{
This function takes input parameters and computes the likelihodd of the observations ocurring based on the input parameters. Returns the Log-likelihood of the observations given the input parameters.  Also contains a switch which allows the usert to force the function to return the estimated probabilities instead of the Log-likelihood.
}
\examples{
TL_Wide(AA_Data,boot=0)
TagLoss(c(0.5,0.1,0.5,0.1), obs=obs, t=t)

}
