% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_wide.R
\name{TL_Wide}
\alias{TL_Wide}
\title{Restructures double-tag-loss data in wide format}
\usage{
TL_Wide(Data, boot = 0)
}
\arguments{
\item{Data}{The input data. Takes a data.frame with oberservations as rows and liberties (first column) and tag categories (subsequent to first) as columns.}

\item{boot}{A nuber value.  Numbers greater than 1 cause the data to be resampled for boot strapping.  Defaults to 0.}
}
\value{
Two input files for use by the DLT function:
     A matrix with tag categories as rows and liberties as columns and a vector of liberties.
}
\description{
Takes data on double tagged data in wide format and re-cofigures it into two input files for the tag-loss model DTL.
}
\examples{
TL_Wide(ab_Data)
}
